`bluetooth`: Remove EXPORT_SYMBOL for check_bt_op
The symbol 'check_bt_op' is no longer needed to be exported as it is not used outside of the current module. This change removes the unnecessary export, simplifying the module and preventing potential misuse.

No functional changes are expected, as the symbol was not accessed by any other modules or components.

Signed-off-by: Robin <hanya8493@gmail.com>